attendance_system/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
sqlalchemy==2.0.20
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
